+++ components/src/main/java/com/mercadolibre/android/andesui/button/factory/AndesButtonAttrsParser.kt
@@ -1,7 +1,6 @@
 package com.mercadolibre.android.andesui.button.factory
 
 import android.content.Context
-import android.graphics.drawable.Drawable
 import android.util.AttributeSet
 import com.mercadolibre.android.andesui.R
 import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonHierarchy
@@ -12,8 +11,8 @@ import com.mercadolibre.android.andesui.button.size.AndesButtonSize
  */
 internal data class AndesButtonAttrs(val andesButtonHierarchy: AndesButtonHierarchy,
                                      val andesButtonSize: AndesButtonSize,
-                                     val andesButtonLeftIcon: Drawable?,
-                                     val andesButtonRightIcon: Drawable?,
+                                     val andesButtonLeftIconPath: String?,
+                                     val andesButtonRightIconPath: String?,
                                      val andesButtonText: String?,
                                      val andesButtonEnabled: Boolean = true)
 
@@ -62,8 +61,8 @@ internal object AndesButtonAttrsParser {
         return AndesButtonAttrs(
                 andesButtonHierarchy = hierarchy,
                 andesButtonSize = size,
-                andesButtonLeftIcon = typedArray.getDrawable(R.styleable.AndesButton_andesButtonLeftIconCustom),
-                andesButtonRightIcon = typedArray.getDrawable(R.styleable.AndesButton_andesButtonRightIconCustom),
+                andesButtonLeftIconPath = typedArray.getString(R.styleable.AndesButton_andesButtonLeftIconPath),
+                andesButtonRightIconPath = typedArray.getString(R.styleable.AndesButton_andesButtonRightIconPath),
                 andesButtonEnabled = typedArray.getBoolean(R.styleable.AndesButton_andesButtonEnabled, true),
                 andesButtonText = typedArray.getString(R.styleable.AndesButton_andesButtonText)
         ).also { typedArray.recycle() }