+++ components/src/main/java/com/mercadolibre/android/andesui/button/hierarchy/AndesButtonIcon.kt
@@ -1,6 +1,5 @@
 package com.mercadolibre.android.andesui.button.hierarchy
 
-import android.graphics.drawable.Drawable
 
 /**
  * Holds data about the icon a button can carry: Icon and Orientation.
@@ -8,7 +7,7 @@ import android.graphics.drawable.Drawable
  * @property icon the icon itself
  * @property orientation an [AndesButtonIconOrientation] that can be [AndesButtonIconOrientation.LEFT] or [AndesButtonIconOrientation.RIGHT]
  */
-class AndesButtonIcon(val icon: Drawable?, private val orientation: AndesButtonIconOrientation) {
+class AndesButtonIcon(val icon: String?, private val orientation: AndesButtonIconOrientation) {
     internal val leftIcon get() = getLeftIcon()
     internal val rightIcon get() = getRightIcon()
 
@@ -17,7 +16,7 @@ class AndesButtonIcon(val icon: Drawable?, private val orientation: AndesButtonI
      *
      * @return the left icon if it exists, null otherwise.
      */
-    private fun getLeftIcon(): Drawable? {
+    private fun getLeftIcon(): String? {
         return when {
             icon != null && orientation == AndesButtonIconOrientation.LEFT -> {
                 icon
@@ -31,7 +30,7 @@ class AndesButtonIcon(val icon: Drawable?, private val orientation: AndesButtonI
      *
      * @return the right icon if it exists, null otherwise.
      */
-    private fun getRightIcon(): Drawable? {
+    private fun getRightIcon(): String? {
         return when {
             icon != null && orientation == AndesButtonIconOrientation.RIGHT -> {
                 icon