+++ components/src/main/java/com/mercadolibre/android/andesui/message/type/AndesMessageTypeInterface.kt
@@ -3,11 +3,11 @@ package com.mercadolibre.android.andesui.message.type
 import android.content.Context
 import android.graphics.drawable.BitmapDrawable
 import android.graphics.drawable.Drawable
-import android.support.v4.content.ContextCompat
 import com.mercadolibre.android.andesui.R
 import com.mercadolibre.android.andesui.button.hierarchy.BackgroundColorConfig
 import com.mercadolibre.android.andesui.color.AndesColor
 import com.mercadolibre.android.andesui.color.toAndesColor
+import com.mercadolibre.android.andesui.icons.OfflineIconProvider
 import com.mercadolibre.android.andesui.message.hierarchy.AndesMessageHierarchyInterface
 import com.mercadolibre.android.andesui.utils.buildColoredCircularShapeWithIconDrawable
 
@@ -32,7 +32,7 @@ internal object AndesNeutralMessageType : AndesMessageTypeInterface() {
     override fun primaryColor() = R.color.andesui_accent_color_500.toAndesColor()
     override fun secondaryColor() = R.color.andesui_accent_color_600.toAndesColor()
     override fun icon(context: Context, hierarchy: AndesMessageHierarchyInterface) = buildColoredCircularShapeWithIconDrawable(
-            ContextCompat.getDrawable(context, R.drawable.andesui_ui_feedback_info_16) as BitmapDrawable,
+            OfflineIconProvider(context).loadIcon("andesui_ui_feedback_info_16") as BitmapDrawable,
             context,
             R.color.andesui_white.toAndesColor(),
             hierarchy.iconBackgroundColor(this)?.colorInt(context),
@@ -64,7 +64,7 @@ internal object AndesSuccessMessageType : AndesMessageTypeInterface() {
     override fun primaryColor() = R.color.andesui_green_500.toAndesColor()
     override fun secondaryColor() = R.color.andesui_green_600.toAndesColor()
     override fun icon(context: Context, hierarchy: AndesMessageHierarchyInterface) = buildColoredCircularShapeWithIconDrawable(
-            ContextCompat.getDrawable(context, R.drawable.andesui_ui_feedback_success_16) as BitmapDrawable,
+            OfflineIconProvider(context).loadIcon("andesui_ui_feedback_success_16") as BitmapDrawable,
             context,
             R.color.andesui_white.toAndesColor(),
             hierarchy.iconBackgroundColor(this)?.colorInt(context),
@@ -96,7 +96,7 @@ internal object AndesWarningMessageType : AndesMessageTypeInterface() {
     override fun primaryColor() = R.color.andesui_orange_500.toAndesColor()
     override fun secondaryColor() = R.color.andesui_orange_600.toAndesColor()
     override fun icon(context: Context, hierarchy: AndesMessageHierarchyInterface) = buildColoredCircularShapeWithIconDrawable(
-            ContextCompat.getDrawable(context, R.drawable.andesui_ui_feedback_warning_16) as BitmapDrawable,
+            OfflineIconProvider(context).loadIcon("andesui_ui_feedback_warning_16") as BitmapDrawable,
             context,
             R.color.andesui_white.toAndesColor(),
             hierarchy.iconBackgroundColor(this)?.colorInt(context),
@@ -128,7 +128,7 @@ internal object AndesErrorMessageType : AndesMessageTypeInterface() {
     override fun primaryColor() = R.color.andesui_red_500.toAndesColor()
     override fun secondaryColor() = R.color.andesui_red_600.toAndesColor()
     override fun icon(context: Context, hierarchy: AndesMessageHierarchyInterface) = buildColoredCircularShapeWithIconDrawable(
-            ContextCompat.getDrawable(context, R.drawable.andesui_ui_feedback_warning_16) as BitmapDrawable,
+            OfflineIconProvider(context).loadIcon("andesui_ui_feedback_warning_16") as BitmapDrawable,
             context,
             R.color.andesui_white.toAndesColor(),
             hierarchy.iconBackgroundColor(this)?.colorInt(context),