+++ components/src/test/java/com/mercadolibre/android/andesui/button/AndesButtonSizeTest.kt
@@ -6,10 +6,10 @@ import android.support.v4.content.ContextCompat
 import com.mercadolibre.android.andesui.BuildConfig
 import com.mercadolibre.android.andesui.R
 import com.mercadolibre.android.andesui.button.factory.IconConfig
+import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonHierarchyInterface
 import com.mercadolibre.android.andesui.button.size.AndesLargeButtonSize
 import com.mercadolibre.android.andesui.button.size.AndesMediumButtonSize
 import com.mercadolibre.android.andesui.button.size.AndesSmallButtonSize
-import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonHierarchyInterface
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
@@ -18,6 +18,8 @@ import org.mockito.Mockito
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
+
+const val ANDESUI_ICON="andesui_icon"
 
 @RunWith(RobolectricTestRunner::class)
 @Config(constants = BuildConfig::class, sdk = [Build.VERSION_CODES.LOLLIPOP])
@@ -68,12 +70,12 @@ class AndesLargeButtonSizeTest {
 
     @Test
     fun `Large button left icon`() {
-        assertEquals(andesLargeButtonSize.iconConfig(hierarchy, icon, null, context), IconConfig(icon, null))
+        assertEquals(andesLargeButtonSize.iconConfig(hierarchy, ANDESUI_ICON, null, context), IconConfig(icon, null))
     }
 
     @Test
     fun `Large button right icon`() {
-        assertEquals(andesLargeButtonSize.iconConfig(hierarchy, null, icon, context), IconConfig(null, icon))
+        assertEquals(andesLargeButtonSize.iconConfig(hierarchy, null, ANDESUI_ICON, context), IconConfig(null, icon))
     }
 
 }
@@ -126,12 +128,12 @@ class AndesMediumButtonSizeTest {
 
     @Test
     fun `Medium button left icon`() {
-        assertNull(andesMediumButtonSize.iconConfig(hierarchy, icon, null, context))
+        assertNull(andesMediumButtonSize.iconConfig(hierarchy, ANDESUI_ICON, null, context))
     }
 
     @Test
     fun `Medium button right icon`() {
-        assertNull(andesMediumButtonSize.iconConfig(hierarchy, null, icon, context))
+        assertNull(andesMediumButtonSize.iconConfig(hierarchy, null, ANDESUI_ICON, context))
     }
 
 }
@@ -185,12 +187,12 @@ class AndesSmallButtonSizeTest {
 
     @Test
     fun `Small button left icon`() {
-        assertNull(andesSmallButtonSize.iconConfig(hierarchy, icon, null, context))
+        assertNull(andesSmallButtonSize.iconConfig(hierarchy, ANDESUI_ICON, null, context))
     }
 
     @Test
     fun `Small button right icon`() {
-        assertNull(andesSmallButtonSize.iconConfig(hierarchy, null, icon, context))
+        assertNull(andesSmallButtonSize.iconConfig(hierarchy, null, ANDESUI_ICON, context))
     }
 
 }