+++ components/src/test/java/com/mercadolibre/android/andesui/button/AndesButtonTest.kt
@@ -1,23 +1,30 @@
 package com.mercadolibre.android.andesui.button
 
 import android.os.Build
-import android.support.v4.content.ContextCompat
+import android.support.constraint.ConstraintLayout
+import com.facebook.common.logging.FLog
+import com.facebook.drawee.backends.pipeline.Fresco
+import com.facebook.imagepipeline.core.ImagePipelineConfig
+import com.facebook.imagepipeline.listener.RequestListener
+import com.facebook.imagepipeline.listener.RequestLoggingListener
+import com.facebook.soloader.SoLoader
 import com.mercadolibre.android.andesui.BuildConfig
 import com.mercadolibre.android.andesui.R
-import com.mercadolibre.android.andesui.button.size.AndesButtonSize
 import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonHierarchy
+import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonIcon
+import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonIconOrientation
+import com.mercadolibre.android.andesui.button.size.AndesButtonSize
 import com.mercadolibre.android.andesui.typeface.getFontOrDefault
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
+import org.junit.Before
+import org.junit.BeforeClass
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
 import org.robolectric.annotation.Config
-import android.support.constraint.ConstraintLayout
-import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonIcon
-import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonIconOrientation
 
 
 @RunWith(RobolectricTestRunner::class)
@@ -25,6 +32,25 @@ import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonIconOrientat
 class AndesButtonTest {
     private var context = RuntimeEnvironment.application
     private lateinit var andesButton: AndesButton
+
+    companion object {
+        @JvmStatic
+        @BeforeClass
+        fun beforeClass() {
+            SoLoader.setInTestMode()
+        }
+    }
+
+    @Before
+    fun setUp() {
+        val requestListeners = setOf<RequestListener>(RequestLoggingListener())
+        val config = ImagePipelineConfig.newBuilder(context)
+                // other setters
+                .setRequestListeners(requestListeners)
+                .build()
+        Fresco.initialize(context, config)
+        FLog.setMinimumLoggingLevel(FLog.VERBOSE)
+    }
 
     @Test
     fun `Only context constructor`() {
@@ -36,7 +62,8 @@ class AndesButtonTest {
         assertEquals(textParams.goneEndMargin, 8)
         assertEquals(andesButton.paddingRight, 16)
         assertEquals(andesButton.paddingLeft, 16)
-        assertEquals(andesButton.textComponent.typeface, context.getFontOrDefault(R.font.andesui_font_semibold))
+        assertEquals(andesButton.textComponent.typeface, context.getFontOrDefault(
+                R.font.andesui_font_semibold))
         assertNull(andesButton.leftIconComponent.drawable)
         assertNull(andesButton.rightIconComponent.drawable)
     }
@@ -58,7 +85,7 @@ class AndesButtonTest {
 
     @Test
     fun `Hierarchy, size and icon constructor`() {
-        val icon = AndesButtonIcon(ContextCompat.getDrawable(context, R.drawable.andesui_icon), AndesButtonIconOrientation.LEFT)
+        val icon = AndesButtonIcon(ANDESUI_ICON, AndesButtonIconOrientation.LEFT)
         andesButton = AndesButton(context, AndesButtonSize.LARGE, AndesButtonHierarchy.TRANSPARENT, icon)
         val textParams = andesButton.textComponent.layoutParams as ConstraintLayout.LayoutParams
         val leftIconParams = andesButton.leftIconComponent.layoutParams as ConstraintLayout.LayoutParams
@@ -76,7 +103,7 @@ class AndesButtonTest {
 
     @Test
     fun `Hierarchy, size and right icon constructor`() {
-        val icon = AndesButtonIcon(ContextCompat.getDrawable(context, R.drawable.andesui_icon), AndesButtonIconOrientation.RIGHT)
+        val icon = AndesButtonIcon(ANDESUI_ICON, AndesButtonIconOrientation.RIGHT)
         andesButton = AndesButton(context, AndesButtonSize.LARGE, AndesButtonHierarchy.TRANSPARENT, icon)
         val textParams = andesButton.textComponent.layoutParams as ConstraintLayout.LayoutParams
         val rightIconParams = andesButton.rightIconComponent.layoutParams as ConstraintLayout.LayoutParams