+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## Fixed
 - Color palette and styles change.
 
+## Changed
+- Icons now are called by string path.
+
 # v1.4.2
 ## Added
 - Test for Andes Message.